+++ project/APMS/classes/wilaya.hpp
@@ -37,8 +37,8 @@ public:
 	
 	void print_wilaya_by_month(const int year,const int& month);
 	void print_wilaya_by_year(const int& year,int choice);
-	int get_wilaya_total_sales_per_year(int year);
-	int get_wilaya_total_sales_per_month(int year,int month);
+	double get_wilaya_total_sales_per_year(int year);
+	double get_wilaya_total_sales_per_month(int year,int month);
 
 
 	void print_wilaya_monthly_farmer_sales(int year,int month);
@@ -49,11 +49,11 @@ public:
 	void print_wilaya_monthly_penalty(int year,int month,string category_name);
 	void print_wilaya_yearly_penalty(int year,string category_name);
 	//functions to get the water and electricity consumption of the wilaya
-	int get_wilaya_yearly_water_consumption(int year);
-	int get_wilaya_yearly_electricity_consumption(int year);
+	double get_wilaya_yearly_water_consumption(int year);
+	double get_wilaya_yearly_electricity_consumption(int year);
 	//functions to get the water and electricity consumption of the wilaya in a month
-	int get_wilaya_monthly_water_consumption(int year,int month);
-	int get_wilaya_monthly_electricity_consumption(int year,int month);
+	double get_wilaya_monthly_water_consumption(int year,int month);
+	double get_wilaya_monthly_electricity_consumption(int year,int month);
 };
